Move `AuthenticatedHTTPClientDecorater` to production
