+++ Sources/Curtz/Authentication API/Token/TokenResponse.swift
@@ -1,0 +1,14 @@
+//
+//  TokenResponse.swift
+//  CurtzUnitTests
+//
+//  Created by George Nyakundi on 04/09/2022.
+//
+
+import Foundation
+
+public struct TokenResponse {
+    let accessToken: String
+    let refreshToken: String
+    let tokenType: String
+}
