+++ Sources/Curtz/Authentication API/Token/TokenService.swift
@@ -1,0 +1,14 @@
+//
+//  TokenService.swift
+//  CurtzUnitTests
+//
+//  Created by George Nyakundi on 04/09/2022.
+//
+
+import Foundation
+
+public protocol TokenService {
+    typealias Result = Swift.Result<String, Error>
+    typealias GetTokenCompletion = (Result) -> Void
+    func getToken(completion: @escaping GetTokenCompletion )
+}
